Set refresh interval to 5m default
Set the default viewer scrape interval to 5m by default. This may
initially appear long, but watchtower is not intended to be a real-time
or event detection tool. A 5m interval allows the app to validate the
targeted environment 288 times a day, which is 288 more times than
our team is checking it today.