Remove config viewer
Removal of the config viewer should result in a much smaller codebase
that accomplishes the same result. Simple modifications the the Detector
struct methods accomplish the same thing as calling the cached
ConfigViewer values, and should generally result in the same number of
API calls being made to the CF API.

As an example, the ConfigViewer was previously set to execute every X
seconds (X=60), and the DriftDetector using that viewer used the same X
value for its interval, just offset by X/2. So eliminating ConfigViewer
and just relying on DriftDetector to make calls to CF ListApps calls
when it needs to is both simpler, and still results in a single call to
ListApps.