Add space detection (#3)
* Add CF Space support to ResourceConfig

* Add error message to cfclient panic output

* Add environment variable expansion support

* Add config, endpoint, and environment variable definitions

* Log cf api being used

* Remove config viewer

Removal of the config viewer should result in a much smaller codebase
that accomplishes the same result. Simple modifications the the Detector
struct methods accomplish the same thing as calling the cached
ConfigViewer values, and should generally result in the same number of
API calls being made to the CF API.

As an example, the ConfigViewer was previously set to execute every X
seconds (X=60), and the DriftDetector using that viewer used the same X
value for its interval, just offset by X/2. So eliminating ConfigViewer
and just relying on DriftDetector to make calls to CF ListApps calls
when it needs to is both simpler, and still results in a single call to
ListApps.

* Modify config file structure

* Add support for disabling app monitoring

* Update readme to be vastly more informative

* Add Space SSH detection capabilities

* Remove unused package dependencies

* Use int instead of float in log output

* Add .gitignore

* Enable passing config file with cli arg

* Add /config endpoint for viewing current config