Create config package to handle Watchtower config
Prior to creating a dedicated package for config handling, the
configuration options for watchtower were beginning to sprawl across
several files, and it became difficult to know where and how
configuration settings were being handled.

With the config package now in place, modifying the watchtower config
is clear and using the Config type is convenient and simple. Many
previously global variables now have much smaller scopes, and
configuration is dealt with via the config file, rather than a mix of
file and command line options.