Add sshMap to AppCache using v2 CAPI
This commit adds an sshMap to the AppCache struct to enable Watchtower
to monitor ssh access on a per-app basis. To do this, I opted to use the
aging v2 cloud controller API. The reason behind this is that the new v3
app structs do not include the "EnableSSH" field. In the v3 api, ssh is
treated as a "feature" and requires sending a request to the CAPI
/v3/GUID/enable_ssh endpoint for each app. I tested a sequential
implementation of this and the code was complex. Parallelizing with
goroutines would have complicated it even more.

SSH information is not available with the go-cfclient, but if the v2 API
is ever deprecated, this section can be rewritten using the v3 API.